requirements.txt
pynput==1.3.10
numpy==1.14.3
pandas==0.23.0
keyboard==0.13.2
opencv_python==3.4.1.15
Pillow==5.2.0
PyQt5==5.11.2
